Add initial version of StringConcatStrategyApp and its tests
This commit introduces a console application that is demonstrating different string concatenation strategies in C#. The goal is to enhance the awareness of performance ramifications these strategies can have in highload systems. Apart from the application itself, several unit tests have been added in order to validate expected behavior of the GreetingService, StringConcatStrategyFactory and UserInputValidator classes.